chore: move CreateDataDisk and ResizeDataDisk to util
